% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_d.R
\name{cutoff_d}
\alias{cutoff_d}
\title{Calculate the maximum vesicle size to be fully sedimented}
\usage{
cutoff_d(t, rcf, rotor, vesicle.density = 1.15, solvent.density = 1,
  viscosity = 0.0155)
}
\arguments{
\item{t}{time (s)}

\item{rcf}{relative centrifugal force}

\item{rotor}{a rotor object}

\item{vesicle.density}{the vesicle density (g/cm^3)}

\item{solvent.density}{the solvent density (g/cm^3)}

\item{viscosity}{of the solvent (g/(cm*s))}
}
\description{
Calculates the maximum vesicle size that is completely sedimented at 
a specified time and density at a specified rcf. Units are noted in parentheses
}

