% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelleted.R
\name{pelleted}
\alias{pelleted}
\title{Calculate the fraction pelleted for a vesicle size}
\usage{
pelleted(t, d, rcf, rotor, vesicle.density = 1.15, solvent.density = 1,
  viscosity = 0.0155)
}
\arguments{
\item{t}{time (s)}

\item{d}{vesicle diameter (nm)}

\item{rcf}{relative centrifugal force}

\item{rotor}{a rotor object}

\item{vesicle.density}{the vesicle density (g/cm^3)}

\item{solvent.density}{the solvent density (g/cm^3)}

\item{viscosity}{of the solvent (g/(cm*s))}
}
\description{
Calculates the fraction pelleted for a vesicle of a specified 
diameter and density when centrifuged at a specified rcf for a specified time. 
Units are noted in parentheses.
}

