% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{rpm2rcf}
\alias{rpm2rcf}
\title{RPM Conversion}
\usage{
rpm2rcf(rpm, rotor)
}
\arguments{
\item{rpm}{velocity of the rotor in rotations-per-minute}

\item{rotor}{a rotor object}
}
\value{
relative centrifugal force (RCF) of the rotor in x g
}
\description{
Converts RPM to RCF for given rotor.
}

