% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{rcf2rpm}
\alias{rcf2rpm}
\title{RCF Conversion}
\usage{
rcf2rpm(rcf, rotor)
}
\arguments{
\item{rcf}{relative centrifugal force in x g}

\item{rotor}{a rotor object}
}
\value{
rotor velocity in rotations-per-minute (RPM)
}
\description{
Converts RCF to RPM for a given rotor.
}

