% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_time.R
\name{sed_time}
\alias{sed_time}
\title{Calculate the time to sediment a vesicle}
\usage{
sed_time(d, rcf, rotor, vesicle.density = 1.15, solvent.density = 1,
  viscosity = 0.0155)
}
\arguments{
\item{d}{vesicle diameter (nm)}

\item{rcf}{relative centrifugal force (x g)}

\item{rotor}{a rotor object}

\item{vesicle.density}{the vesicle density (g/cm^3)}

\item{solvent.density}{the solvent density (g/cm^3)}

\item{viscosity}{of the solvent (g/(cm*s))}
}
\description{
Calculates the time needed for complete sedimentation of a specified
size and density at a specified rcf. Units are noted in parentheses.
}

